+++ crates/miden-testing/src/tx_context/builder.rs
@@ -290,43 +290,10 @@ impl TransactionContextBuilder {
         // merkle paths of assets and storage maps, in order to test lazy loading.
         // Otherwise, load the full account.
         let tx_inputs = if self.is_lazy_loading_enabled {
-            let (account, block_header, partial_blockchain, input_notes) = tx_inputs.into_parts();
-            // Construct a partial vault that tracks the empty word, but none of the assets
-            // that are actually in the asset tree. That way, the partial vault has the same
-            // root as the full vault, but will not add any relevant merkle paths to the
-            // merkle store, which will test lazy loading of assets.
+            let (_account, block_header, partial_blockchain, input_notes) = tx_inputs.into_parts();
             // Note that we use self.account instead of account, because we cannot do the same
             // operation on a partial vault.
-            let mut partial_vault = PartialVault::default();
-            partial_vault.add(self.account.vault().open(Word::empty()).into())?;
-
-            // Construct a partial storage that tracks the empty word in all storage maps, but none
-            // of the other keys, following the same rationale as the partial vault above.
-            let storage_header = self.account.storage().to_header();
-            let storage_maps =
-                self.account.storage().slots().iter().filter_map(
-                    |storage_slot| match storage_slot {
-                        StorageSlot::Map(storage_map) => {
-                            let mut partial_storage_map = PartialStorageMap::default();
-                            partial_storage_map
-                                .add(storage_map.open(&Word::empty()))
-                                .expect("adding the first proof should never error");
-                            Some(partial_storage_map)
-                        },
-                        _ => None,
-                    },
-                );
-            let partial_storage = PartialStorage::new(storage_header, storage_maps)
-                .expect("provided storage maps should match storage header");
-
-            let account = PartialAccount::new(
-                account.id(),
-                account.nonce(),
-                account.code().clone(),
-                partial_storage,
-                partial_vault,
-                None,
-            )?;
+            let account = minimal_partial_account(&self.account)?;
 
             TransactionInputs::new(account, block_header, partial_blockchain, input_notes)?
         } else {
@@ -386,4 +353,42 @@ impl Default for TransactionContextBuilder {
     fn default() -> Self {
         Self::with_existing_mock_account()
     }
+}
+
+/// Creates a minimal [`PartialAccount`] from the provided full [`Account`].
+fn minimal_partial_account(account: &Account) -> anyhow::Result<PartialAccount> {
+    // Construct a partial vault that tracks the empty word, but none of the assets
+    // that are actually in the asset tree. That way, the partial vault has the same
+    // root as the full vault, but will not add any relevant merkle paths to the
+    // merkle store, which will test lazy loading of assets.
+    let mut partial_vault = PartialVault::default();
+    partial_vault.add(account.vault().open(Word::empty()).into())?;
+
+    // Construct a partial storage that tracks the empty word in all storage maps, but none
+    // of the other keys, following the same rationale as the partial vault above.
+    let storage_header = account.storage().to_header();
+    let storage_maps =
+        account.storage().slots().iter().filter_map(|storage_slot| match storage_slot {
+            StorageSlot::Map(storage_map) => {
+                let mut partial_storage_map = PartialStorageMap::default();
+                partial_storage_map
+                    .add(storage_map.open(&Word::empty()))
+                    .expect("adding the first proof should never error");
+                Some(partial_storage_map)
+            },
+            _ => None,
+        });
+    let partial_storage = PartialStorage::new(storage_header, storage_maps)
+        .expect("provided storage maps should match storage header");
+
+    let account = PartialAccount::new(
+        account.id(),
+        account.nonce(),
+        account.code().clone(),
+        partial_storage,
+        partial_vault,
+        None,
+    )?;
+
+    Ok(account)
 }